optimize lexer code in all layers

+++ src/main/java/io/github/ititus/commons/math/statistics/DurationStatistics.java
@@ -33,7 +33,7 @@ public final class DurationStatistics {
         return averagingMode.average(durations);
     }
 
-    public long count() {
+    public int count() {
         return durations.size();
     }
