+++ src/main/java/io/github/ititus/commons/math/statistics/IntegerStatistics.java
@@ -73,10 +73,9 @@ public final class IntegerStatistics {
         return averagingMode.average(numbers);
     }
 
-    public long count() {
+    public int count() {
         return numbers.size();
     }
-
 
     public BigInteger min() {
         return Collections.min(numbers);
