+++ src/main/java/io/github/ititus/commons/math/statistics/RationalStatistics.java
@@ -72,10 +72,9 @@ public final class RationalStatistics {
         return averagingMode.average(numbers);
     }
 
-    public long count() {
+    public int count() {
         return numbers.size();
     }
-
 
     public BigRational min() {
         return Collections.min(numbers);
